feat(medication): Use `BigDecimal` for medication stock
The `stock` property in the `Medication` domain model and related entities and mappers has been changed from `Float?` to `BigDecimal?`.

This change provides more precise representation of medication stock values.

Database schema version 1 has been updated to reflect this change, storing `stock` as `TEXT` instead of `REAL`.

The `isValidDose()` extension function for `String` now also uses `BigDecimal` for validation.
The `AddMedicationViewModel` has been updated to parse the stock input as `BigDecimal`.
Preview data in `MedicineCabinetScreen` has been updated to use `BigDecimal`.

+++ core/domain/src/main/kotlin/io/floriax/medschedule/core/domain/model/Medication.kt
@@ -1,5 +1,6 @@
 package io.floriax.medschedule.core.domain.model
 
+import java.math.BigDecimal
 import java.time.Instant
 
 /**
@@ -11,7 +12,7 @@ import java.time.Instant
 data class Medication(
     val id: Long = 0,
     val name: String,
-    val stock: Float?,
+    val stock: BigDecimal?,
     val doseUnit: String,
     val notes: String,
     val createdAt: Instant = Instant.now()
